// frontend/src/api/generated/models/action-choice.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 行動選択肢
 * @export
 * @interface ActionChoice
 */
export interface ActionChoice {
  /**
   * 表示する選択肢のテキスト
   * @type {string}
   * @memberof ActionChoice
   */
  text: string
  /**
   * 行動のタイプ（move, investigate, interact等）
   * @type {string}
   * @memberof ActionChoice
   */
  action_type: string
  /**
   *
   * @type {string}
   * @memberof ActionChoice
   */
  description?: string | null
  /**
   *
   * @type {object}
   * @memberof ActionChoice
   */
  metadata?: object | null
}